+++ src/pages/MyInformation/components/Designs/index.tsx
@@ -1,9 +1,6 @@
-import { FAILED_TO_GET_MY_DESIGNS } from 'constants/errors';
-
 import { List, Typography } from '@material-ui/core';
-import { useCommonSnackbar } from 'components/CommonSnackbar/useCommonSnackbar';
 import EmptyContent from 'dumbs/EmptyContent';
-import { useGetMyDesigns } from 'pages/MyInformation/hooks/useGetMyDesigns';
+import { useMyDesigns } from 'pages/MyInformation/hooks/useMyDesigns';
 import { tabItemLengthAtom } from 'pages/MyInformation/recoils';
 import { useEffect } from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
@@ -12,50 +9,26 @@ import styled from 'styled-components';
 import { theme } from 'themes';
 import { DEFAULT_LIST_LENGTH } from 'utils/requestType';
 
-import { DesignItemResponse } from '../../hooks/types';
 import DesignItem from '../DesignItem';
 import { useRenderEmptyContent } from '../InformationTabs/useRenderEmptyContent';
 
 const Designs = (): React.ReactElement => {
-  const { data, error, size, setSize } = useGetMyDesigns();
-  const setTabItemLength = useSetRecoilState(tabItemLengthAtom);
+  const { isLoading, hasLastCursor, designs, roadMore } = useMyDesigns();
   const emptyContent = useRenderEmptyContent();
 
-  useCommonSnackbar({
-    message: FAILED_TO_GET_MY_DESIGNS,
-    severity: 'error',
-    dependencies: [error],
-  });
+  const setTabItemLength = useSetRecoilState(tabItemLengthAtom);
 
-  const getResponseDesigns = () =>
-    data?.reduce(
-      (designs: DesignItemResponse[], { payload }) => designs.concat(payload),
-      [],
-    );
-  const isLoading = data == null;
-  const designs = isLoading
-    ? [...Array(DEFAULT_LIST_LENGTH)]
-    : getResponseDesigns() ?? [];
   const isEmpty = !isLoading && designs.length === 0;
-  const lastData = data?.[data.length - 1];
-  const hasLastCursor =
-    (lastData?.payload ?? []).length > 0 && lastData?.meta.last_cursor != null;
 
   useEffect(() => {
     setTabItemLength(designs.length);
   }, [designs]);
 
-  const getNextDesigns = (): void => {
-    if (hasLastCursor) {
-      setSize(size + 1);
-    }
-  };
-
   return (
     <StyledList>
       <InfiniteScroll
         dataLength={designs.length}
-        next={getNextDesigns}
+        next={roadMore}
         hasMore={hasLastCursor}
         loader={
           <Loader>
@@ -69,7 +42,7 @@ const Designs = (): React.ReactElement => {
 
             return (
               <DesignItem
-                isLoading={data == null}
+                isLoading={isLoading}
                 key={index}
                 {...design}
                 showDivider={showDivider}
